Make getters on incoming request and socket readonly properties
Readonly is enforced by typescript but in js these values are
mutable. You can also escape the ts compiler to write to these
properties. This implementation is more accurate to nodejs which
does allow the request to be overwritten. This functionality is
relied upon by express. If we don't support express then do
we even fake ttp?