Force proxy and increase response timeout to 30 seconds
The timeout of 100ms was too low. unix has a 30 second timeout
on tcp and node imposes no timeout on http socket server. Setting
to 30 seems reasonable even if it should be overkill for loopback
like connectivity.

Forcing the proxy all the time. Disables some security feature
but I didn't look into what it was. Without it we get misses of
some files. This needs to be debugged further and a better explaination
as to why force was added in the first place.