feat: mutable request (#2)
Certain properties of the request / socket need to be mutable. They are
in node and libraries like express rely on this functionality to rewrite
parts of the request object.

Removed the proxy settings. Found it causing issues. Need to revisit it
in the future but do not recommend using it for now.

---------

Co-authored-by: Sepehr Laal <5657848+3p3r@users.noreply.github.com>